+++ src/main/java/org/overbaard/review/tool/security/github/GitHubCredential.java
@@ -22,7 +22,7 @@ public class GitHubCredential {
         return tokenHeader;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return gitHubUser.getId();
     }
